feat(sdk): refactor code structure
