+++ packages/sdk/src/utils/index.ts
@@ -1,5 +1,9 @@
 import { Provider } from 'fuels';
 
+export * from './wallet';
+export * from './errors';
+export * from './domain';
+
 export const getTxParams = (provider: Provider) => {
   const gasConfig = provider.getGasConfig();
   return {
@@ -7,22 +11,3 @@ export const getTxParams = (provider: Provider) => {
     gasLimit: 1_000_000
   };
 };
-
-export class InvalidDomainError extends Error {
-  constructor() {
-    super('Invalid domain characters.');
-    this.name = 'InvalidDomainError';
-  }
-}
-
-export class NotFoundBalanceError extends Error {
-  constructor() {
-    super('Balance not found.');
-    this.name = 'NotFoundBalanceError';
-  }
-}
-
-export const isValidDomain = (domain: string) => /^@?[a-zA-Z0-9_]+$/.test(domain);
-
-export * from './wallet';
-export * from './price';
